+++ src/main/java/tornadofx/Nodes.kt
@@ -1188,6 +1188,18 @@ fun ComboBoxBase<*>.editableWhen(predicate: ObservableValue<Boolean>) = apply {
     editableProperty().bind(predicate)
 }
 
+fun TableView<*>.editableWhen(predicate: ObservableValue<Boolean>) = apply {
+    editableProperty().bind(predicate)
+}
+
+fun TreeTableView<*>.editableWhen(predicate: ObservableValue<Boolean>) = apply {
+    editableProperty().bind(predicate)
+}
+
+fun ListView<*>.editableWhen(predicate: ObservableValue<Boolean>) = apply {
+    editableProperty().bind(predicate)
+}
+
 /**
  * This extension function will make sure that the given [onHover] function will always be calles
  * when ever the hoverProperty of the given node changes.